// backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TypeCompte = {
  Cash: 'Cash',
  Banque: 'Banque'
} as const

export type TypeCompte = (typeof TypeCompte)[keyof typeof TypeCompte]


export const TypeTransaction = {
  DEPENSE: 'DEPENSE',
  REVENU: 'REVENU',
  TRANSFER: 'TRANSFER'
} as const

export type TypeTransaction = (typeof TypeTransaction)[keyof typeof TypeTransaction]